backend/tests/load/requirements.txt
locust==2.17.0
requests==2.31.0
sseclient-py==1.8.0
